Add LICENSE and README to the addon
Following the submission docs[1]. This ensures that users of the plugin
have these files, and in the case of the LICENSE helps them to comply
with the license.

Replace the top-level README.md with a symbolic link to the copy in the
addon, so that we do not need to keep the two in sync. GitHub displays
this correctly at the top level of the repo.

Sadly GitHub's LICENSE file handling mishandles symbolic links, so we
duplicate the LICENSE file.

[1] https://docs.godotengine.org/en/latest/community/asset_library/submitting_to_assetlib.html