Don't suppress class categories containing only non-property blocks
Previously, if the property list for a class was empty, then the
category generated for that class would not be used, even if there were
other specialised blocks for the class. It just so happens that at
present that case can't arise, but in a future commit it will.

Instead, show the category if it contains any blocks at all.