entry block: Allow multi-line opening statement
It may be that there isn't a 1-1 correspondance between blocks and lines
of code. For a contrived example, you might want to wrap the parameter
list, or include a documentation comment. And in particular, a future
commit will add a preamble inside the generated function for an entry
block that converts a parameter to a different type. Or you could
imagine wanting to generate a top-level variable or signal to accompany
an entry block.