category_factory: Warn if class doesn't have expected property
Previously, if the plugin developer incorrectly specified that a
property block should be created for a class that doesn't have a
property by that name, property_to_blocklist(null) would be called,
raising a type error since that function accepts a non-nullable
Dictionary.

This condition represents a bug, but to ease debugging we can log a
warning and continue.