parameter_input: Warn & proceed if Types.cast() fails
Previously, if the node snapped into a parameter had a different type
than its parent block, code generation would fail with the obscure
error:

> Trying to return value of type "Nil" from a function which the return
> type is "String".

This is because Types.cast() can return null if no cast is possible; but
get_string()'s return type is annotated as String; and null is not a
String.

Instead, check Types.can_cast() before calling cast(). In the case where
the cast would return null, log a warning with some useful information
and then return the generated parameter directly. With luck, Godot will
be able to implicitly convert; and it's better than generating invalid
code.