Add ‘On [body] entered/exited’ blocks for Area2D and RigidBody2D
Each block ‘compiles’ to a callback function, which needs to be
connected to the signal in question. To do this, add an additional
code-generation step that generates an _init() function to connect
signals.

A more idiomatic approach would be to connect to the signals in the
_ready() function. This would involve merging additional code with the
user-supplied _ready() function, so is deferred for now. _init() works
fine.