Pong demo: Add area for spawning balls
In Pong the balls spawn at a random position from the center vertical
line. To avoid the maths, one option is to pick a random point from an
Area2D with a thin vertical rectangle as collision shape. The area has
group "ball_spawn_area" so it can be referenced.