WIP: Add named classes for final blocks
Instead of defining blocks in category_factory.gd, define blocks ahead
of time, each in their own GDScript file. Although it is not a concept
in GDScript itself, we can consider block classes such as EntryBlock to
be abstract block types.

By doing this, we are able to reduce duplication in the block script
resource associated with each BlockCode node. In particular, the
serialized data can simply refer to "PrintBlock", and all of the
EntryBlock properties expected for the print block are implied.

This commit only adds ReadyBlock and PrintBlock as an example, and
likely breaks everything else.