Fixe base class identification
For previous classes with custom blocks, like SimpleCharacter, you had
to define a method that specified the base class so the picker would
know which blocks to populate itself with.

However, there is a function script.get_instance_base_type() that gets
this automatically from whichever class the script extends. Much nicer
than having to define a function for this every time.

This change also fixes an error when opening the pong game.