block_code/picker: Store the AudioStreamPlayer in the VAR_DICT
The new generated AudioStreamPlayer was held by a local variable, then
added as a child of the node originally.

It is okay for only one AudioStreamPlayer in the node. However, some
nodes may have mulitiple AudioStreamPlayers, then the local variable
from the code blocks will be conflicted, due to the same variable name.

Therefore, store the AudioStreamPlayer into the node's global VAR_DICT
with a key passed from the parameter "name" directly. This also avoids
searching in the node tree.

https://phabricator.endlessm.com/T35501