block_category_display: Make categories buttons to show/hide blocks
When all the categories are shown, you may need to do quite a bit of
scrolling to find the block you want or just to discover what's there.
This changes the category display so that each category is a button that
can show or hide its blocks. All blocks are initially hidden to be able
to see all the available categories.

The expander is implemented as a flat button with a icon. Godot does
provide the Tree class with similar functionality, but it's a bit
complex to use. More importantly TreeItems are their own class that
don't derive from Node. That would break the current Block usage. The
Tree arrow icons are reused, though. To give a little more indication
that the category can be expanded, the color is lightened relative to
the expanded state.

https://phabricator.endlessm.com/T35507