Reset snap point variables in DragManager.drag_end
In the `drag_end` function, make sure to reset `previewing_snap_point`
in addition to `preview_block`. This ensures that block snapping will
work correctly if the same block is dragged a second time. In addition,
defer cleaning up preview_block and previewing_snap_point to the end of
the function. This avoids some confusion related to nested if
statements.

https://phabricator.endlessm.com/T35524