Category factory: Rename signals to methods
The rest of the blocks match the GDScript nomenclature. Except for a few
"simple" or higher level abstractions.

This also removes the signal manager. Keep the await for scene ready in
a separate block so it can be used explicitly when needed. Add it to the
Lifecycle category.

Keep the category "Signal" for now as they need overhaul in ordering and
colors.