Add template for upload to asset library
I overlooked the section of the documentation for
https://github.com/deep-entertainment/godot-asset-lib-action which says
that you have to provide a template.

Personally I don't see why a built-in template couldn't be provided: all
fields in the request are optional and default to their existing values;
and if you are using this action you presumably want the normal
GitHub-based export. But this way we can keep the description etc. in
version control.

https://phabricator.endlessm.com/T35503