Add plugin version to exported plugin.cfg
This boils down to a call to `git describe`.

For an unreleased snapshot this currently looks like: v0.2.0-6-gc38af04.

When run against a commit that matches an annotated (not lightweight)
tag, it will be the name of that tag. (v0.2.0 is annotated; v0.1.0 is
lightweight. Signed tags are always annotated.)

https://phabricator.endlessm.com/T35503