Scratch-like category UI
This adds all uncollapsed block categories back into one large scroll container, but now with buttons that can navigate to each category.
This allows the user to see the entire selection of blocks, without the tedium of scrolling.
I think this will be nicer for adding in a little menu to create variables within the scrollbox as well.