main_panel: Correct tooltip for block picker toggle button
The tooltip previously claimed that the shortcut was Ctrl + Space, but in fact
it is Ctrl + Backslash.

Fix this error; and adjust the tooltip style to match the tooltip on the
equivalent button in the script editor view. I personally don't like the
camel-casing of BackSlash but that's what Godot calls that key.