Duplicate instead of modifying block code resources from other scenes
When editing a block code resource in the context of a scene, make sure
we are saving the result in the same scene by creating a new copy. This
improves the workflow when editing BlockCode nodes from instantiated
child scenes: a user is unlikely to expect a change with the Block Code
editor in one scene to modify a different scene file.

https://phabricator.endlessm.com/T35541