Add a feature to override BlockCode from a child scene
This reorganizes the empty state widgets in BlockCanvas, adding a new
set of actions when the user selects a node which contains block code,
but the block code can't be edited. This commit also includes changes
to better handle when a BlockCode node is inside a node from a different
scene but with editable children.

https://phabricator.endlessm.com/T35541