Make SimpleCharacter more flexible
To allow different movement kinds for different game possibilities. The
previous is now the default and named "as top-down". In addition, now
"as platformer" and "as spaceship" are available.

The speed is now exposed as a property to prevent the block getting too
long with 3 parameters. The speed vector is used differently in each
situation:
- for platformer, Y is the jump speed
- for spaceship, X is the rotation speed

For platformer, the gravity is read from project settings, but also
there is a method to update it.

Remove the gravity custom setting from the project setting, that was
added back at some point by mistake.

https://phabricator.endlessm.com/T35549