Select parent node when a BlockCode is selected
In order for a node to be selected on the 2D canvas while its associated
BlockCode is being edited, automatically select the current BlockCode
node's parent. This requires some changes in BlockCodePlugin to deal
with cases when multiple nodes are selected in the Scene panel and in
the Inspector panel.

https://phabricator.endlessm.com/T35541