Add version to block script data
In the plugin, add a new constant for the current data version. This
will start in 0 and be bumped each time the data format changes.

In the BlockScriptData resource, add an integer version field which
defaults to zero. The version will be set to the current version when
the resource is saved.

The actual migration is TBD. Only a warning is printed for now.

https://phabricator.endlessm.com/T35547