Move script generation to InstructionTree
Handling the script generation in the UI canvas code seems misplaced,
and it depends on nothing in the UI except for the current canvas
window. Move the generation to InstructionTree using an interface taking
an array of Node. This will allow testing the code generation without
using the block canvas.