Polyfill Node.is_part_of_edited_scene()
Godot 4.3+ exposes `Node.is_part_of_edited_scene()` to GDScript, which is
analogous to `Engine.is_editor_hint()` but determines whether the `@tool`
script's node is itself being edited, as opposed to instantiated elsewhere in
the editor.

Reimplement this function in GDScript.

https://phabricator.endlessm.com/T35540

Move node_is_part_of_edited_scene to new Util module

This will allow it to be reused in other scenes which have the same problem as
tooltip did.