Don't recreate resources when editing plugin scenes
Previously, opening and saving parameter_block.tscn (for example) in the editor
would cause its StyleBoxFlat resource to be replaced by an identical
sub_resource with a new unique identifier. This is because its _ready()
function runs when the scene is being edited, and previously would
unconditionally re-set the stylebox.

This either causes needless diff churn, or requires extra work by the developer
to discard that hunk.

In each case, guard the change with `node_is_part_of_edited_scene`, and
remove it from the saved scene.