Parameter input: change text modification logic
Previously any text change was sending the modified signal, adding too
many actions to the undo/redo history (one per typed character).
This was also capturing shortcuts like Ctrl-Z for undoing and sending
the modified signal, thus piling to the history, not going back.

Instead, send the modified signal when the line edit is submitted (enter
key pressed) or when it loses focus and the text has changed.

Do this for the 3 line edit nodes.

https://phabricator.endlessm.com/T35566