Fix inspector plugin leak
Handling the inspector plugin as a static variable of the BlockCode
class means it can never be removed from inspector since we never know
when all BlockCode instances have been destroyed. That means it had to
be leaked as shown in the verbose Godot output:

WARNING: ObjectDB instances leaked at exit (run with --verbose for details).
     at: cleanup (core/object/object.cpp:2209)
Leaked instance: EditorPlugin:1833129019089 - Node name:

Instead, add it once from the plugin entry point. That means it can also
be removed when the plugin exits rather than being leaked. This follows
the pattern documented in the Inspector Plugin tutorial[1].

1. https://docs.godotengine.org/en/stable/tutorials/plugins/editor/inspector_plugins.html.