Plugin: Only skip duplicate work when BlockCode nodes selected
In dbd7643, an early return was added to avoid updating the UI when the
same BlockCode node has been selected. When a non-BlockCode has been
selected, the main panel needs to be told about it so that it can show
some friendly messages about adding a BlockCode node. However, any
non-BlockCode node is represented as null, which means that changes to a
selected non-BlockCode node can't be differentiated. Skip the early
return in that case.

https://phabricator.endlessm.com/T35572