Move DragManager.get_tree_scope to InstructionTree
This static function has nothing to do with the DragManager class since
it's simply walking a tree of Nodes and looking for a Block that has a
non-empty scope property set. I'm not sure it really fits in
InstructionTree either, but that's a better fit at least. This avoids
some dependency issues between DragManager, its inner Drag class, and
BlockCanvas.

https://phabricator.endlessm.com/T35494