Build Simple node scenes in code
The method to instantiate a scene internally in 6d017d0 is clever, but
the real classes and internal classes create a circular dependency that
cause import errors. Since there are only a couple child nodes that are
significantly updated by the exported variables, just create the whole
scene in code.

There should be no functional changes besides dropping the internal
scenes. The only change is that the simple nodes can only be added as a
node type and not by the scene file. I don't think that will be missed.

https://phabricator.endlessm.com/T35494