Remove class_name from all other internal scripts
Except for:
- The resources: they must be classes
- Block and subclasses: they are accessed by class using the ClassDB
- SnapPoint: there is an "is SnapPoint" check
- the CategoryFactory (bug in Godot 4.2.1)

With this, the class is not registered by name in ClassDB and it can't
be referenced directly in other scripts, it must be preloaded. Also, if
the script extends Node or a subclass of Node, it won't appear in the UI
to the end user when adding a node.

Another advantage: it won't collide with classes defined by the end user
in the project. Our current Types class has a good chance to conflict
given the generic name.

Also: delete PackedSceneTreeNode and PackedScene abandoned
resources.