CategoryFactory: Add defaults to all blocks
All possible blocks gain default values. So they can be used directly
without having to type a value. And most importantly, so the script is
not generated with syntax errors.

The defaults are opinionated, they have to be. Eg. set scale to double
size, change position so the node2d moves to the right-bottom.

Also: remove "change by" block for modulate and visible properties. They
don't make sense.

https://phabricator.endlessm.com/T35562