Block: Allow to grab focus
Set the focus mode on Blocks so they can process input events. When
dragged on the canvas, set the dropped block to be focused. Note the
subtle interaction between FocusMode.FOCUS_ALL, which allows focusing by
mouse click, and MouseFilter.MOUSE_FILTER_IGNORE, which filters out
mouse click events. The combination means that the mouse click will be
handled in Godot for focusing the block, but the mouse click can't be
processed in _gui_input().