Refresh block script when Block Code panel is shown
It is possible for the list of animations to change based on changes
done in another panel. This causes the list of animation value blocks in
the Block Code panel to become stale. Ideally we should watch the
mixer_updated signal in an AnimationPlayer node, but at the moment this
would be difficult to implement. Watching for the panel being shown will
work around the issue for a reasonable number of situations.

https://phabricator.endlessm.com/T35564