Dynamically add blocks for variables
Using the new block definition. These are the only blocks that are not
stored in the catalog. Instead, they are dynamically generated when the
BlockScriptSerialization changes.

Also: Move constant for the builtin category properties to the
constants.gd script.

Also: Finally remove the preloaded scene blocks from the
category_factory.gd script. The single place where those scenes are now
instantiated is the ui/util.gd script.

https://phabricator.endlessm.com/T35591