BlockScriptSerialization: Add block library and instantiation
This changes BlockScriptSerialization to manage its own list of
available block definitions and instantiate them, replacing the existing
functionality that was in CategoryFactory and ui/util.gd. The change
makes it easier to understand the state of the block catalog, because it
is a component of the block script itself instead of spread across
different static variables.

In addition, this adds a singleton block editor context, which keeps
track of the current block code node and its associated block script.
The context object provides enough information for different UI
components to update automatically when the block script changes.

https://phabricator.endlessm.com/T35564