SimpleCharacter: Don't impose a collision shape
Change the logic so there is no fallback 100x100 collision shape if a
texture is not provided. This way the user can customize their character
by adding their own sprites and collision shapes as node children. The
simple setup of passing a texture is still maintained. Although it has
been found to be inflexible (can't resize the texture from the canvas,
can't change the shape of the collision).

Addresses #215