ParameterInput: Add a separate option data property
Instead of inferring that there is a list of options based on the
current value, ParameterInput expects to be given a list of options,
and get_raw_input and set_raw_input understand that the raw input is a
value from the list.

In addition, change TemplateEditor itself to determine the list of
options for a given parameter. This option data is always from the
parameter_defaults property, regardless of the block's current value.

https://phabricator.endlessm.com/T35564