TemplateEditor: Remove the OPTION parameter type
Instead, we infer that a parameter has a set of options by the existence
of an OptionData as its default value.

In addition, add a way for a code template to include a plain, unquoted
string by using `{{parameter}}`. Previously this happened automatically
for any parameter whose value was an OptionData.

https://phabricator.endlessm.com/T35564