Stop changing the editor selection
Back in a808f67 and 43dd08b, the editor selection was dynamically
updated so that both the BlockCode node and its parent were selected.
This would open the parent node in the 2D/3D scene editor while the
BlockCode node was open in the block canvas.

However, selecting both nodes at the same time puts the editor into a
multi-selection mode that makes interaction with the scene tree dock
awkward. For example, clicking on another node does not select it for
dragging. Furthermore, since some parts of the code operate on the
selection while others wait to see the inspector's edited object is,
there can be strange interactions when the selection is changed in the
middle of an input action.

What we really want is a way to say "make this the active scene in the
scene editor" without selecting the node like we can with BlockCode
nodes and the block canvas. That doesn't appear possible with Godot, so
just always look at what the currently edited object in the inspector is
and stop trying to activate the BlockCode parent in the scene editor.

https://phabricator.endlessm.com/T35648