blocks: Add groups option block extension
Several blocks work with node groups. Rather than having the user enter
the group, offer the current groups as options. This is trickier than
expected since Godot offers no API to get the current groups. Finding
them involves 2 parts:

* Get scene local groups by calling get_groups() on each node in the
  scene.

* Get project global groups (starting in 4.3) by getting the
  `global_group` values in the project settings.

Monitoring for changes to the global groups is possible by listening for
project settings changes. Monitoring for local group changes is not
possible since the only events that occur are internal to Godot and not
available in GDScript.

https://phabricator.endlessm.com/T35647