SimpleSpawner: Allow spawning nodes as child of scene
In the current implementation, spawned nodes are children of the spawner. In
some cases this behaviour is what you want, but it means it's not possible to
change the spawn position of future scenes without moving already-spawned
scenes.

Add a new property to SimpleSpawner which allows choosing a new mode where
spawned scenes are made children of the root scene.

The change in spawn_start() is necessary because, if the Start Spawning block
is used in the On Ready entrypoint, this results in spawn_start() being called
from the SimpleSpawner's _ready() function. At this point, the SimpleSpawner
itself is ready but the root scene is not. In this case, attempting to add a
child to the root scene logs an error, and a later attempt to get the spawned
scene's parent to remove it returns undefined.