SimpleSpawner: Rename frequency to period
This property is the time, in seconds, to wait between spawning scenes. This is
a period/interval/wavelength, not a frequency, which would be measured in Hz,
or 1/seconds.

Adjust the name of the property and its documentation. Remove the unnecessary
do_set_spawn_frequency() method in favour of generating a property assignment.
Update the example scene accordingly.

The instances of “frequency” that I have not changed are in the example scene's
on-screen documentation. In this case it was actually used correctly: the
keyboard action to increase the frequency reduces the property which is now
called period, and the decrease action respectively increases the period.