simple_nodes: Add SimpleEnding node
This provides a label that displays either a win or lose message along
with blocks to set or clear the label during the game. A LabelSettings
instance is provided to easily customize the text settings for the
label.

Since the Label isn't created until simple_setup and that runs after
_ready, a signal is used to defer the game_over and reset methods until
the label exists. Otherwise, using the blocks from a "when starting"
block would crash.

https://phabricator.endlessm.com/T35661