simple_nodes: Add SimpleEnding node
This provides a label that displays either a win or lose message along
with blocks to set or clear the label during the game. Label is
inherited directly so that all the label properties are available in the
inspector.

We want the label to be empty when the scene is run, but there may be
text persisted in the scene file. When not in the editor, the text is
cleared when entering the scene tree. Normally this would be done in
_ready, but it's likely that gets replaced by a block script.

https://phabricator.endlessm.com/T35661