Add the concept of a hidden block definition
This makes it possible to define a block that cannot be added to new block
programs, but which will still work if present in a block program.

The intended use is to fix SimpleSpawner's "spawn frequency" property and
blocks, which is in fact the spawn period (frequency is the reciprocal of
period).