Script Window: Don't clamp font color alpha channel to 0.65
Because the CodeEdit's readonly property is set, TextEdit forces the alpha
channel of all text rendered to be no higher than the alpha channel of the
default colour for text in disabled controls, which is 0.65 as I type this.

While we do not want the code in this popover to be editable, we do want
it to be legible.

Override the relevant font color on the control to set its alpha channel to 1,
effectively disabling this undesired clamping.