SimpleSpawner: Add spawn_frequency migration
Renaming the spawn_frequency properties and blocks will break any
existing scene since those are serialized. Migrate the spawn_frequency
property by providing it as an additional unstored property that simply
wraps spawn_period. Migrate the block names and arguments using the
serialization renaming support from the previous commit. A test using a
scene created before the renaming is provided to verify the migration.

While this does not handle the worst compatibility breakage, it can be
used as a model for other migrations. Note that the migration stub in
MainPanel is likely too late for any legitimate fixes. Any migration
likely needs to happen when the scene is loaded and instantiated, which
is before MainPanel gets a hold of the objects.

https://phabricator.endlessm.com/T35547