serialization: Handle renamed blocks and arguments
If a block or its arguments has been renamed, loading a scene with the
old block name will fail and the user will be required to replace the
block. Handle renamed blocks and arguments by migrating the names to the
new values in the serialization resources. In order for the setters to
run in the editor, the classes need to be tool scripts.

https://phabricator.endlessm.com/T35547