serialization: Handle renamed blocks
If a block has been renamed, loading a scene with the old block name
will fail and the user will be required to replace the block. Handle
renamed blocks by setting the name to the new value in the serialization
resources. In order for the setter to run in the editor, the classes
need to be tool scripts.

https://phabricator.endlessm.com/T35547