Add tool to regenerate POT file
Currently the only way to regenerate a POT file is to go through the
Regenerate POT dialog in the Localization tab of the Project Settings.
This adds a Project->Tools menu item that drives the dialog from code
using our POT file path. Hat tip to KoBeWi for the procedure. Hopefully
in the future we'll have a Godot CLI interface to do this.