Add tool to update POT files
The list of translated files to include in the POT file is stored in the
project settings. The primary way to update it is through the POT
Generation dialog when a new file is added, but we know that we
generally want to add all of the plugin's files for translation. This
adds a Project->Tools menu item that finds all the relevant files and
updates the project setting.