locale: Create common message catalogs
Create message catalogs for all the locales supported in the Godot
editor except for es_AR and pt_BR. Due to a bug in Godot's locale
matching[1], these would be preferred over the country-less variants for
all countries. For example, the pt_BR translation would be used for
pt_PT when the generic pt translation would be preferred.

The catalogs were created like so:

```
msginit --no-translator -l $locale -i godot_block_coding.pot -o $locale.po
```

The catalogs are also added to the project settings despite that setting
not being used in the editor. Nodes that are used in games such as the
simple nodes can still get translations via the main domain this way.

1. https://github.com/godotengine/godot/issues/90677