BlockCanvas: Drag & drop the node's set property block from Inspector if CTRL is pressed
Drag & drop the node's set property block from Inspector if the modifier
key CTRL is pressed. To detect pressing CTRL key, hook the input event
and set the _modifier_ctrl flag with the CTRL key's pressed state. If
the _modifier_ctrl is true, then drop a set Variable block. Otherwise,
it should be a get Variable block of the property.

https://phabricator.endlessm.com/T35649