ci: Use -s not --script to run tests
gut's cmdline script accepts a number of command-line arguments. It
appears that the Godot engine removes most of its own options from the
argument list available to user code, but it leaves any `--script FOO`
or `-s FOO` arguments in place.

gut 9.2.1 understands both the `--script` and `-s` forms, but gut 9.3.0
only understands the `-s` form and raises an error if the `--script`
form is used.

https://github.com/bitwes/Gut/issues/667