scripts: Add regenerate pot CLI script and wrapper
This can be used to execute TxUtils.regenerate_pot_file() from the CLI.
The GDScript is a SceneTree script passed via Godot's --script CLI
option. The shell script is a wrapper to execute Godot with all the
necessary options.

Since TxUtils.regenerate_pot_file() works by using the editor's
Localization panel, the script has to be run with --editor. That causes
a bunch of leaks that we can't do anything about. The script tries to
soothe the user about the pile of warnings and errors at the end.