Add Godot wrapper script
To ease running Godot from the command line, provide a wrapper script
that looks for common Godot executables. As an escape hatch, the `GODOT`
environment variable can be used to set exactly how Godot should be
executed. This is very Linux-centric with only `godot` or the Godot
Flatpak supported, but it can be extended as needed.