Blocks background: Simplify outline drawing
Remove the separate stroke line "edge_polygon" which was used only for
control blocks. Since this line is connected to the "stroke_polygon" it
can be joined with it. Also the "edge_polygon" was adding a redundant
draw to non-control blocks.

Also remove the outline_middle calculation. Taking screenshots as proof,
this extra calculation has no effect in the drawing.